begin adding geometries and materials to Chuck
have decoupled (mostly) the geometry and material scenegraph nodes from the actual rendering implementation.
Realizing now that the scenegraph needs to be entirely decoupled from any GPU orgraphics implementation, managing cpu-side data only. The renderer (which should live as a Chugin or otherwise independent from chuck core) is the one that processes this information and does the graphics calls to make it happen

Material.SetLocalUniforms is still coupled however. Need to somehow have a way to express the value, names, types, and layout of arbitrary uniform data. somethign similar to a buffer + vertex buffer layout (but CPU side) should do the trick.